Create simon_says.py
Game: Simon says
Do you know the game "Simon says"?
The instructor, "Simon", says what the players should do, e.g. "jump in the air" or "close your eyes".
The players should follow the instructions only if the phrase starts with or end with "Simon says",
If it starts or ends with the words "Simon says", the players say: "I + (what you would do)!"
Otherwise, the players say: "I won't do it!".

This is a practice to use "find" function and slice a string.